mpOpenAPI: refactor MergeServerXMLTest away
Move the tests from MergeServerXMLTest into MergeConfigServerXMLTest.

MergeConfigServerXMLTest now includes equivalents to all the tests from
MergeConfigTest, plus the additional tests specific to server.xml config
previously in MergeServerXMLTest.

Having all the tests in one class makes it easier to share setup code
and reduces the number of server restarts required.
